chore: Fixes minor code structure issues
